old-tests/submission/Microsoft/async/async_006.htm

<!DOCTYPE html>
<html>
    <head>
        <title>A dynamically created external script executes asynchronously</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures a dynamically created external script executes asynchronously." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dev.w3.org/html5/spec/scripting-1.html#force-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">        
        
  
        var t = async_test("dynamically created external script executes asynchronously");
  
        function timeout()
        {
            t.step(function(){ assert_equals(document.getElementById("testresult").innerHTML, "321")});
            t.done();
        }
  
        var timer = setTimeout(timeout, 8000);
        
        // This test uses a script (log.py) to change script order execution
        // See async_003.htm and async_004.htm for how this works
        // In this test 
        //  * The third script tag below should execute first
        //  * Then second 'two' dynamically created script tag should execute since it only has a one second delay
        //  * Then first 'one' dynamically created script tag should execute since it has a three second delay
        // 
        //  Thus the assert above is to check that the innerHTML == 321
        //
        function log(text) 
        {
            var textNode = document.createTextNode(text);
            document.getElementById("testresult").appendChild(textNode);
        }
        
        </script>
        
        <span style="display : none" id="testresult"></span>
        <script type="text/javascript">
            var one = document.createElement("script");
            one.src="log.py?sec=3&id=1";
            document.head.appendChild(one);

            var two = document.createElement("script");
            two.src="log.py?sec=1&id=2";
            document.head.appendChild(two);
        </script>
        <script type="text/javascript">
            log('3');
        </script>
    </body>
</html>
